Adapt tests
(cherry picked from commit 6dff8f7)
